bus_optimization_pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "bus-transit-simulation-pipeline"
description = "A comprehensive bus transit simulation pipeline with Azure cloud deployment capabilities"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "UA3Innovations", email = "ua3innovations@gmail.com"}
]
maintainers = [
    {name = "UA3Innovations", email = "ua3innovations@gmail.com"}
]
keywords = ["simulation", "transit", "bus", "transportation", "azure", "cloud", "optimization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Scientific/Engineering :: Visualization",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dynamic = ["version"]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "PyYAML>=5.4.0",
    "python-dateutil>=2.8.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]
azure = [
    "azure-containerinstance>=2.0.0",
    "azure-storage-blob>=12.0.0",
    "azure-identity>=1.7.0",
]
all = [
    "bus-transit-simulation-pipeline[dev,azure]"
]

[project.urls]
Homepage = "https://github.com/UA3Innovations/graduation_project"
"Bug Reports" = "https://github.com/UA3Innovations/graduation_project/issues"
Source = "https://github.com/UA3Innovations/graduation_project"
Documentation = "https://github.com/UA3Innovations/graduation_project#readme"

[project.scripts]
bus-simulation = "pipeline.pipeline_runner:main"
bus-deploy = "scripts.deploy_to_azure:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["data/*.csv", "config/*.yaml"]

[tool.setuptools_scm]
write_to = "src/_version.py"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers --disable-warnings --cov=src"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true 